test: split optimize-imports transform coverage
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
